pyproject.toml
[project]
name = "ragit"
version = "0.1.0"
description = "RAGIT - RAG with Gateway-Backend Architecture"
authors = [
    {name = "GyuChul Team"}
]
dependencies = [
    # Backend dependencies
    "fastapi>=0.109.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "psycopg2-binary==2.9.9",
    "alembic==1.12.1",
    "pydantic[email]>=2.7.4",
    "python-multipart>=0.0.7",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-decouple==3.8",

    # Frontend dependencies
    "nicegui==1.4.24",
    "requests>=2.31.0",

    # Gateway dependencies
    "httpx==0.25.2",

    # Worker dependencies
    "amqp==5.3.1",
    "billiard==4.2.1",
    "celery==5.5.3",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1.2",
    "click-repl==0.3.0",
    "colorama==0.4.6",
    "kombu==5.5.4",
    "packaging>=23.2,<25",
    "prompt-toolkit==3.0.52",
    "python-dateutil==2.9.0.post0",
    "redis==6.4.0",
    "six==1.17.0",
    "tzdata==2025.2",
    "vine==5.1.0",
    "wcwidth==0.2.13",


    # Vector DB & RAG dependencies
    "pymilvus>=2.4.0",
    "langchain-milvus>=0.1.0",
    "langchain-huggingface>=0.1.0",
    "langchain-core>=0.3.0",
    "rank-bm25>=0.2.2",
    "torch>=2.0.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",

    # Git History Tracking
    "GitPython>=3.1.0",
    "tree-sitter>=0.20.0,<0.21.0",
    "tree-sitter-languages==1.8.0",
    "diff-match-patch>=20230430",

    # LLM & Utilities
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",

    # SDK dependencies
    "psutil>=5.9.0",
]
requires-python = ">=3.11"
readme = "README.md"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend", "frontend", "gateway", "rag_worker", "ragit_sdk"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[project.scripts]
# Main RAGIT CLI (SDK based)
ragit = "ragit_sdk.cli:main"

# Legacy compatibility is no longer needed as start.py/stop.py have been removed

# Direct SDK access
ragit-process = "ragit_sdk.process_manager:main"
ragit-docker = "ragit_sdk.docker_manager:main"

[project.urls]
Homepage = "https://github.com/your-repo/RAGIT"
Documentation = "https://github.com/your-repo/RAGIT/docs"
Repository = "https://github.com/your-repo/RAGIT.git"
Issues = "https://github.com/your-repo/RAGIT/issues"